# backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Electronic Educare API",
    "description": "API documentation for the Electronic Educare backend.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/subject/fetch": {
      "get": {
        "tags": [
          "Subjects"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/subject/add": {
      "post": {
        "tags": [
          "Subjects"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "subjectName": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "code": {
                    "example": "any"
                  },
                  "classId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/timetable/": {
      "post": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "classId": {
                    "example": "any"
                  },
                  "sectionId": {
                    "example": "any"
                  },
                  "academicYearId": {
                    "example": "any"
                  },
                  "entries": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "parameters": [
          {
            "name": "classId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sectionId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/uploads/cloudinary/single": {
      "post": {
        "tags": [
          "Uploads"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "folder": {
                    "example": "any"
                  },
                  "tags": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/uploads/cloudinary/bulk": {
      "post": {
        "tags": [
          "Uploads"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "folder": {
                    "example": "any"
                  },
                  "tags": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/fetch": {
      "get": {
        "tags": [
          "Subjects"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/add": {
      "post": {
        "tags": [
          "Subjects"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "example": "any"
                  },
                  "subject": {
                    "example": "any"
                  },
                  "class": {
                    "example": "any"
                  },
                  "marks": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  },
                  "dueDate": {
                    "example": "any"
                  },
                  "courseName": {
                    "example": "any"
                  },
                  "courseCode": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "duration": {
                    "example": "any"
                  },
                  "credits": {
                    "example": "any"
                  },
                  "department": {
                    "example": "any"
                  },
                  "prerequisites": {
                    "example": "any"
                  },
                  "instructor": {
                    "example": "any"
                  },
                  "startingDate": {
                    "example": "any"
                  },
                  "venue": {
                    "example": "any"
                  },
                  "date": {
                    "example": "any"
                  },
                  "time": {
                    "example": "any"
                  },
                  "noOfStudents": {
                    "example": "any"
                  },
                  "role": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "schoolName": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  },
                  "consent": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "rating": {
                    "example": "any"
                  },
                  "feedback": {
                    "example": "any"
                  },
                  "schoolId": {
                    "example": "any"
                  },
                  "stream": {
                    "example": "any"
                  },
                  "fees": {
                    "example": "any"
                  },
                  "programType": {
                    "example": "any"
                  },
                  "programLabel": {
                    "example": "any"
                  },
                  "totalStudents": {
                    "example": "any"
                  },
                  "installments": {
                    "example": "any"
                  },
                  "desc": {
                    "example": "any"
                  },
                  "subjectName": {
                    "example": "any"
                  },
                  "code": {
                    "example": "any"
                  },
                  "classId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/cloudinary/single": {
      "post": {
        "tags": [
          "Uploads"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "folder": {
                    "example": "any"
                  },
                  "tags": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cloudinary/bulk": {
      "post": {
        "tags": [
          "Uploads"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "folder": {
                    "example": "any"
                  },
                  "tags": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}